requirements.txt
distro==1.9.0
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabula-py==2.9.3
tzdata==2024.1
